improve IsColorTerminal windows support
