+++ go.mod
@@ -5,13 +5,14 @@ go 1.16
 require (
 	github.com/fatih/color v1.16.0
 	github.com/mattn/go-colorable v0.1.13
-	github.com/mattn/go-isatty v0.0.20
+	golang.org/x/sys v0.14.0
+	golang.org/x/term v0.14.0
 )
 
 require (
-	// Only used for test/benchmark/comparision.
+	// Only used for test/benchmark/comparison.
 	github.com/nwidger/jsoncolor v0.3.2
-	// Only used for test/benchmark/comparision.
+	// Only used for test/benchmark/comparison.
 	github.com/segmentio/encoding v0.3.6
 	github.com/stretchr/testify v1.8.4
 )
